pyproject.toml
[project]
name = "vaultkeeper"
version = "0.1.0"
description = "Obsidian vault MCP server with semantic search, Bases support, and intelligent query routing"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0.0",
    "chromadb>=0.5.0",
    "watchfiles>=0.24.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1.0",
    "tiktoken>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.5.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/vaultkeeper"]

[tool.ruff]
line-length = 100
target-version = "py311"
